Remove role code from requests
