+++ src/app/services/actions.service.ts
@@ -30,13 +30,6 @@ export class ActionsService {
   public findWidgets (devices, location: ILocation): Array<CloudDevice> {
     return devices.filter(device => +device.location === +location.id);
   }
-  public DeleteRole (role: IRole) {
-    if (confirm( this.translate.get('Are you sure to delete this role?')['value'])) {
-      this.requests.deleteRole(role.id);
-      this.notification.InvokeRoleDelete(role);
-      this.router.navigateByUrl('/roles');
-    }
-  }
 
   public scrollTop () {
     const el = new SimpleBar(document.getElementById('layout'));
